+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.3] - 2022-03-07
 
 ### Added
 - Project `curifactory` command tests
+
+### Changed
+- Project init .gitignore handling to check for/add a blank line before adding
+  the curfiactory section
 
 ### Fixed
 - Notebook experiment folder not being created on init and from experiment.